+++ Trainer_v5/Trainer.Source/Helpers.cs
@@ -7,7 +7,7 @@ namespace Trainer_v5
 	public static class Helpers
 	{
 		public static bool IsGameLoaded => GameSettings.Instance != null && HUD.Instance != null;
-		public static string Version => "5.2.5";
+		public static string Version => "5.2.6";
 		public static string TrainerVersion => $"Trainer v{Version}";
 		public static bool IsDebug => false;
 		public static string DiscordUrl => "https://discord.com/invite/J584aG";
@@ -73,6 +73,7 @@ namespace Trainer_v5
 			{"DisableForcePause", false},
 			{"DisableForceFreeze", false},
 			{"AutoAcceptHostingDeals", false},
+			{"AutoMaxMarketShare", false},
 			{"Experimental", false},
 		};